# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Condominium Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings/{buildingId}/units:
    get:
      operationId: getBuildingUnits
      tags: [units]
      summary: List units for a building
      parameters:
        - $ref: "#/components/parameters/buildingId"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, all]
          description: Filter by status (default active)
      responses:
        "200":
          description: List of units with ownership info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnitSummary"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit (apartment) in a building
      parameters:
        - $ref: "#/components/parameters/buildingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  # ── Units ─────────────────────────────────────────────────────────────────
  /units/{unitId}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get unit details with ownership and tenancy
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Unit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitDetail"
        "404":
          description: Not found
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update unit details
      parameters:
        - $ref: "#/components/parameters/unitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{unitId}/archive:
    post:
      operationId: archiveUnit
      tags: [units]
      summary: Archive a unit (preserves all history)
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Unit archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{unitId}/restore:
    post:
      operationId: restoreUnit
      tags: [units]
      summary: Restore an archived unit to active status
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Unit restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
...
components:
  parameters:
...
    buildingId:
      name: buildingId
      in: path
      required: true
      schema:
        type: string
    unitId:
      name: unitId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Building:
      type: object
      required: [id, companyId, name, locality, status, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        locality:
          type: string
        postcode:
          type: ["string", "null"]
        country:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UnitInput:
      type: object
      required: [unitNumber]
      properties:
        unitNumber:
          type: string
          minLength: 1
          description: e.g. "Apt 1", "Unit 4A"
        unitType:
          type: string
          enum: [apartment, garage, commercial, other]
          default: apartment
        floor:
          type: ["integer", "null"]
    UnitUpdate:
      type: object
      properties:
        unitNumber:
          type: string
          minLength: 1
        unitType:
          type: string
          enum: [apartment, garage, commercial, other]
        floor:
          type: ["integer", "null"]
    Unit:
      type: object
      required: [id, companyId, buildingId, unitNumber, unitType, status, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        buildingId:
          type: string
        unitNumber:
          type: string
        unitType:
          type: string
          enum: [apartment, garage, commercial, other]
        floor:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, archived]
        activatedAt:
          type: ["string", "null"]
          format: date-time
        archivedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UnitSummary:
      allOf:
        - $ref: "#/components/schemas/Unit"
        - type: object
          properties:
            ownerName:
              type: ["string", "null"]
            ownerEmail:
              type: ["string", "null"]
            ownerStatus:
              type: ["string", "null"]
              enum: [active, pending, null]
            tenantName:
              type: ["string", "null"]
            tenantStatus:
              type: ["string", "null"]
              enum: [active, pending, null]
    UnitDetail:
      allOf:
        - $ref: "#/components/schemas/Unit"
        - type: object
          required: [building]
          properties:
            building:
              $ref: "#/components/schemas/Building"
            owner:
              $ref: "#/components/schemas/UnitMembership"
            tenant:
              $ref: "#/components/schemas/UnitMembership"
...
    UnitMembership:
      type: object
      required: [id, unitId, role, status, invitedName, invitedEmail, createdAt]
      properties:
        id:
          type: string
        unitId:
          type: string
        userId:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, tenant]
        status:
          type: string
          enum: [pending, active, revoked]
        invitedName:
          type: string
        invitedEmail:
          type: string
        invitationExpiresAt:
          type: ["string", "null"]
          format: date-time
        activatedAt:
          type: ["string", "null"]
          format: date-time
        revokedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...